Create base and test for Parser and LoadLocalFile
LoadLocalFile is a function that receive a path to
a file and return the data of said file in string
format. If the file doenst exist or the path isnt
provided, it return an Error.

Parser is a function that receive a string and will
decodify the data in it returning a list of games
found in the given string with a description of the
game. If it not receive a string, will return null.