Add new atribute to parser and create a script to show the result
The ranking atribute is now returned to the user when it request the
list of games.

In the index.js, is now a simple script that load a local file and
pass it data to the Parser and request the parsed information.
With the parsed data, it shows to the user, via console.log

Obs.: The script to shows the data is not client-side optimized.